Add evaluations in course variable in tests
